requirements.txt
# 小红书生成大师 - Python 依赖

# 核心依赖
playwright>=1.40.0    # 浏览器渲染（图片生成）
pyyaml>=6.0           # YAML 解析（Markdown 头部）

# 可选依赖
# requests>=2.28.0  # 用于 API 调用
# colorama>=0.4.6   # 用于彩色输出
